Improve activemq cache sync logic.
