+++ components/active-mq-support/org.wso2.carbon.cache.sync.active.mq/src/main/java/org/wso2/carbon/cache/sync/active/mq/manager/internal/CacheSyncActiveMQManagerDataHolder.java
@@ -19,7 +19,7 @@
 package org.wso2.carbon.cache.sync.active.mq.manager.internal;
 
 /**
- * Data holder for the ActiveMQCacheManager.
+ * Data holder for the ActiveMQ cache manager.
  */
 public class CacheSyncActiveMQManagerDataHolder {
